Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/backend/requirements.txt
Flask==1.1.2
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo==2.3.0

MarkupSafe==2.0.1
jinja2==3.0.3
itsdangerous==2.0.1
werkzeug==2.0.3

flask_opentracing
flask_cors
prometheus_client
prometheus-flask-exporter

jaeger-client
opentelemetry-api
opentelemetry-sdk
opentelemetry.instrumentation
opentelemetry.instrumentation.flask
opentelemetry.instrumentation.requests
